change structure of StatusEffectInventory
actors are no longer spawned for each status effect. Instead, the status effects are done by StatusEffectInventory itself to reduce the amount of Actors at any given time. StatusEffectManager handles all actions for adding and removing status effects, and executing callback functions when added, on a timer, or when removed.