tweaks to nerf AM
Chain now miss step of 5 instead of 300, and step range of 300 instead of 600. First Damage also 150 instead of 160.
RuneFlurryProj Damage knocked back down to 15 (giving 60 rather than 80 damage on a hit). It was upped when the magnet damage was decreased, but I think that was a mistake. Too strong.
RuneGuard now upped to 10 secs instead of 8 secs. A slight boost amongst the nerfs.